chore: remove design pattern factory to simplify the codebase and refactor the project test : adapt the changes above and merge transform test with extract one.
